Use macro for all shims
