+++ crates/containerd-shim-wasmer/Cargo.toml
@@ -3,13 +3,14 @@ name = "containerd-shim-wasmer"
 version.workspace = true
 edition.workspace = true
 
+[[bin]]
+name = "containerd-shim-wasmer-v1"
+path = "src/main.rs"
+
 [dependencies]
 anyhow = { workspace = true }
-containerd-shim = { workspace = true }
 containerd-shim-wasm = { workspace = true }
 log = { workspace = true }
-oci-spec = { workspace = true, features = ["runtime"] }
-ttrpc = { workspace = true }
 
 tokio = "1.32.0"
 wasmer = { version = "4.1.2" }
@@ -19,15 +20,3 @@ wasmer-wasix = { version = "0.12.0" }
 [dev-dependencies]
 containerd-shim-wasm-test = { workspace = true }
 serial_test = "*"
-
-[[bin]]
-name = "containerd-shim-wasmer-v1"
-path = "src/bin/containerd-shim-wasmer-v1/main.rs"
-
-[[bin]]
-name = "containerd-shim-wasmerd-v1"
-path = "src/bin/containerd-shim-wasmerd-v1/main.rs"
-
-[[bin]]
-name = "containerd-wasmerd"
-path = "src/bin/containerd-wasmerd/main.rs"
