+++ crates/containerd-shim-wasmer/src/tests.rs
@@ -5,7 +5,7 @@ use containerd_shim_wasm_test::modules::*;
 use containerd_shim_wasm_test::WasiTest;
 use serial_test::serial;
 
-use crate::instance::WasmerInstance as WasiInstance;
+use crate::WasmerInstance as WasiInstance;
 
 #[test]
 #[serial]
